Add tests
useNodeBadge is a bit too gnarly to mock, so the code has been pulled
into a new file
