+++ src/composables/node/useNodeBadge.ts
@@ -2,9 +2,10 @@ import _ from 'es-toolkit/compat'
 import { computed, onMounted, watch } from 'vue'
 
 import { useNodePricing } from '@/composables/node/useNodePricing'
+import { usePriceBadge } from '@/composables/node/usePriceBadge'
 import { useComputedWithWidgetWatch } from '@/composables/node/useWatchWidget'
 import { BadgePosition, LGraphBadge } from '@/lib/litegraph/src/litegraph'
-import type { LGraph, LGraphNode } from '@/lib/litegraph/src/litegraph'
+import type { LGraphNode } from '@/lib/litegraph/src/litegraph'
 import { useSettingStore } from '@/platform/settings/settingStore'
 import { app } from '@/scripts/app'
 import { useExtensionStore } from '@/stores/extensionStore'
@@ -12,7 +13,6 @@ import type { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
 import { useColorPaletteStore } from '@/stores/workspace/colorPaletteStore'
 import { NodeBadgeMode } from '@/types/nodeSource'
-import { adjustColor } from '@/utils/colorUtil'
 
 /**
  * Add LGraphBadge to LGraphNode based on settings.
@@ -27,6 +27,7 @@ export const useNodeBadge = () => {
   const settingStore = useSettingStore()
   const extensionStore = useExtensionStore()
   const colorPaletteStore = useColorPaletteStore()
+  const priceBadge = usePriceBadge()
 
   const nodeSourceBadgeMode = computed(
     () =>
@@ -45,29 +46,6 @@ export const useNodeBadge = () => {
   const showApiPricingBadge = computed(() =>
     settingStore.get('Comfy.NodeBadge.ShowApiPricing')
   )
-  function getCreditsBadge(price: string): LGraphBadge {
-    const isLightTheme = colorPaletteStore.completedActivePalette.light_theme
-    return new LGraphBadge({
-      text: price,
-      iconOptions: {
-        unicode: '\ue96b',
-        fontFamily: 'PrimeIcons',
-        color: isLightTheme
-          ? adjustColor('#FABC25', { lightness: 0.5 })
-          : '#FABC25',
-        bgColor: isLightTheme
-          ? adjustColor('#654020', { lightness: 0.5 })
-          : '#654020',
-        fontSize: 8
-      },
-      fgColor:
-        colorPaletteStore.completedActivePalette.colors.litegraph_base
-          .BADGE_FG_COLOR,
-      bgColor: isLightTheme
-        ? adjustColor('#8D6932', { lightness: 0.5 })
-        : '#8D6932'
-    })
-  }
 
   watch(
     [
@@ -141,7 +119,7 @@ export const useNodeBadge = () => {
           let creditsBadge
           const createBadge = () => {
             const price = nodePricing.getNodeDisplayPrice(node)
-            return getCreditsBadge(price)
+            return priceBadge.getCreditsBadge(price)
           }
 
           if (hasDynamicPricing) {
@@ -169,50 +147,18 @@ export const useNodeBadge = () => {
           'litegraph:set-graph',
           () => {
             for (const node of app.canvas.graph?.nodes ?? [])
-              updateSubgraphCredits(node)
+              priceBadge.updateSubgraphCredits(node)
           }
         )
         app.canvas.canvas.addEventListener<'subgraph-converted'>(
           'subgraph-converted',
-          (e) => updateSubgraphCredits(e.detail.subgraphNode)
+          (e) => priceBadge.updateSubgraphCredits(e.detail.subgraphNode)
         )
       },
       afterConfigureGraph() {
         for (const node of app.canvas.graph?.nodes ?? [])
-          updateSubgraphCredits(node)
+          priceBadge.updateSubgraphCredits(node)
       }
     })
   })
-  function updateSubgraphCredits(node: LGraphNode) {
-    if (!node.isSubgraphNode()) return
-    node.badges = node.badges.filter((b) => !isCreditsBadge(b))
-    const newBadges = collectCreditsBadges(node.subgraph)
-    if (newBadges.length > 1) {
-      node.badges.push(getCreditsBadge(newBadges.length + ' x Partner Nodes'))
-    } else {
-      node.badges.push(...newBadges)
-    }
-  }
-  function collectCreditsBadges(
-    graph: LGraph,
-    visited: Set<string> = new Set()
-  ): (LGraphBadge | (() => LGraphBadge))[] {
-    if (visited.has(graph.id)) return []
-    visited.add(graph.id)
-    const badges = []
-    for (const node of graph.nodes) {
-      badges.push(
-        ...(node.isSubgraphNode()
-          ? collectCreditsBadges(node.subgraph, visited)
-          : node.badges.filter((b) => isCreditsBadge(b)))
-      )
-    }
-    return badges
-  }
-
-  function isCreditsBadge(badge: LGraphBadge | (() => LGraphBadge)): boolean {
-    return (
-      (typeof badge === 'function' ? badge() : badge).icon?.unicode === '\ue96b'
-    )
-  }
 }
